+++ CHANGELOG.md
@@ -23,6 +23,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### New features
 
+* The template language now supports the same string pattern system as revsets
+  (e.g. as used in the `diff_contains` revset).
+
+  This has been wired up as `string.matches` such that
+  `"foo".matches(regex:'[a-f]o+')` now works.
+
 ### Fixed bugs
 
 ### Packaging changes